Adding input handlers and such. Did way too much for one commit. Refactor very necessary. Boing

+++ TRIWAV/Assets/Scripts/Constants.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class Constants {
-	public const double PERFECT = 0.033;
-	public const double GREAT = 0.092;
-	public const double GOOD = 0.142;
-	public const double POOR = 0.225;
+	public const float PERFECT = 0.033f;
+	public const float GREAT = 0.092f;
+	public const float GOOD = 0.142f;
+	public const float POOR = 0.225f;
 }
